add windows support
now windows losers can use atoment

+++ src/interpreter.cc
@@ -4,9 +4,9 @@
 #include "keywords.hh"
 #include "util.hh"
 
-ATM::Arglist createArguments(size_t &i, vector <Lexer::Token> &tokens, ATM::Language_Components &atm) {
+ATM::Arglist createArguments(uint32_t &i, vector <Lexer::Token> &tokens, ATM::Language_Components &atm) {
 	ATM::Arglist arglist;
-	size_t j = i + 1;
+	uint32_t j = i + 1;
 	for (; tokens[j].type != Lexer::TokenType::EndOfArguments; ++j) {
 		if (tokens[j].value.length() > 0)
 		switch (tokens[j].type) {
@@ -15,7 +15,7 @@ ATM::Arglist createArguments(size_t &i, vector <Lexer::Token> &tokens, ATM::Lang
 					arglist.push_back(stoi(tokens[j].value));
 				}
 				catch (const std::invalid_argument& ia) {
-					printf("Invalid argument at line %ld col %ld\n", tokens[j].line, tokens[j].column);
+					printf("Invalid argument at line %d col %d\n", (int)tokens[j].line, (int)tokens[j].column);
 					exit(1);
 				}
 				break;
@@ -33,7 +33,7 @@ ATM::Arglist createArguments(size_t &i, vector <Lexer::Token> &tokens, ATM::Lang
 					arglist.push_back(ATM_Pointer(atm.variables[tokens[j].value]));
 				}
 				catch (const std::invalid_argument& ia) {
-					printf("Invalid argument at line %ld col %ld\n", tokens[j].line, tokens[j].column);
+					printf("Invalid argument at line %d col %d\n", (int)tokens[j].line, (int)tokens[j].column);
 					exit(1);
 				}
 				break;
@@ -43,13 +43,13 @@ ATM::Arglist createArguments(size_t &i, vector <Lexer::Token> &tokens, ATM::Lang
 					arglist.push_back(atm.stack[atm.variables[tokens[j].value]]);
 				}
 				catch (const std::invalid_argument& ia) {
-					printf("Invalid argument at line %ld col %ld\n", tokens[j].line, tokens[j].column);
+					printf("Invalid argument at line %d col %d\n", (int)tokens[j].line, (int)tokens[j].column);
 					exit(1);
 				}
 				break;
 			}
 			default: {
-				printf("Error: unexpected token %d in arguments list (index %ld) at line %ld col %ld\n", (int)tokens[i].type, j, tokens[j].line, tokens[j].column);
+				printf("Error: unexpected token %d in arguments list (index %d) at line %d col %d\n", (int)tokens[i].type, j, (int)tokens[j].line, (int)tokens[j].column);
 			}
 		}
 	}
@@ -60,12 +60,12 @@ ATM::Arglist createArguments(size_t &i, vector <Lexer::Token> &tokens, ATM::Lang
 }
 
 void interpretTokens(vector <Lexer::Token> tokens, ATM::Language_Components& atm) {
-	for (size_t i = 0; i<tokens.size(); ++i) {
+	for (uint32_t i = 0; i<tokens.size(); ++i) {
 		atm.variables["MEM_SIZE"] = atm.stack.size();
 		switch (tokens[i].type) {
 			case Lexer::TokenType::Function: {
 				if (atm.functions[tokens[i].value] == nullptr) {
-					printf("Function %s not found at line %ld col %ld\n", tokens[i].value.c_str(), tokens[i].line, tokens[i].column);
+					printf("Function %s not found at line %d col %d\n", tokens[i].value.c_str(), (int)tokens[i].line, (int)tokens[i].column);
 					exit(1);
 				}
 				string function = tokens[i].value; // save it because we modify i later
@@ -89,7 +89,7 @@ void interpretTokens(vector <Lexer::Token> tokens, ATM::Language_Components& atm
 				break;
 			}
 			default: {
-				printf("Error: Unexpected token %d at line %ld col %ld\n", (int)tokens[i].type, tokens[i].line, tokens[i].column);
+				printf("Error: Unexpected token %d at line %d col %d\n", (int)tokens[i].type, (int)tokens[i].line, (int)tokens[i].column);
 				exit(1);
 			}
 		}
