+++ src/keywords.cc
@@ -3,15 +3,15 @@
 
 void keyword_def(ATM::Language_Components& atm, ATM::Arglist args) {
 	if (args.size() != 2) {
-		printf("keyword_def: requires 2 arguments (given %ld)\n", args.size());
+		printf("keyword_def: requires 2 arguments (given %d)\n", (int)args.size());
 		exit(1);
 	}
 	if (args[0].index() != 0) {
-		printf("keyword_def: first argument must be a string (given %ld)\n", args[0].index());
+		printf("keyword_def: first argument must be a string (given %d)\n", (int)args[0].index());
 		exit(1);
 	}
 	if ((args[1].index() != 1) && (args[1].index() != 2)) {
-		printf("keyword_def: second argument must be an integer or a pointer (given %ld)\n", args[1].index());
+		printf("keyword_def: second argument must be an integer or a pointer (given %d)\n", (int)args[1].index());
 		exit(1);
 	}
 	if (args[1].index() == 1) atm.variables[get <string> (args[0])] = get <ATM_Integer> (args[1]);
@@ -20,7 +20,7 @@ void keyword_def(ATM::Language_Components& atm, ATM::Arglist args) {
 
 void keyword_delete(ATM::Language_Components& atm, ATM::Arglist args) {
 	if (args.size() != 1) {
-		printf("keyword_delete: requires 1 argument (given %ld)\n", args.size());
+		printf("keyword_delete: requires 1 argument (given %d)\n", (int)args.size());
 		exit(1);
 	}
 	try {
