+++ src/lexer.hh
@@ -16,7 +16,7 @@ namespace Lexer {
 	struct Token {
 		TokenType type;
 		string    value;
-		size_t    line, column;
+		uint32_t  line, column;
 		Token(TokenType type, string value, size_t ln, size_t col) : type(type), value(value), line(ln), column(col) {}
 	};
 	vector <Token> tokenize(string source);
